Improve system call wrappers
This change improves copy_file_range(), sendfile(), splice(), openpty(),
closefrom(), close_range(), fadvise() and posix_fadvise() in addition to
writing tests that confirm things like errno and seeking behavior across
platforms. We now less aggressively polyfill behavior with some of these
functions when the platform support isn't available. Please see:

https://justine.lol/cosmopolitan/functions.html

+++ libc/sysv/consts/fio.h
@@ -4,11 +4,11 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const uint64_t FIOASYNC;
-extern const uint64_t FIOCLEX;
-extern const uint64_t FIONBIO;
-extern const uint64_t FIONCLEX;
-extern const uint64_t FIONREAD;
+extern const uint32_t FIOASYNC;
+extern const uint32_t FIOCLEX;
+extern const uint32_t FIONBIO;
+extern const uint32_t FIONCLEX;
+extern const uint32_t FIONREAD;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
